ASSIGNMENT - 1 [ Smart Home - Alarm System ! ]
Alarm-System:
Arduino Uno alarm that detects movement, alerts by LED when some movement is observed.

Abstract
The need for home security alarm systems nowadays is a serious demand. As the number of crimes are increasing every day, there has to be something that will keep us safe. We are all aware of the high end security systems present in the market but they are not easily available to everyone. We therefore intend to provide a solution by constructing a cost efficient electronic system that has the capability of sensing the motion of the intruders and setting off the alarm. The basic idea behind this project is that all the bodies generate some heat energy in the form of infrared which is invisible to human eyes. But, it can be detected by electronic motion sensor.

The project involves the use of Arduino, motion sensor, buzzer, LED and a simple program. The sensor detect any motion in its permissible range and triggers the alarm.

It will also send the signal to Arduino which processes the signal and set off the alarm along with detection by indicating via RED LED and BUZZER ! . With this system we can easily set up a security alarm in our home for unwanted intruders.

Reference Link to view my Full Project in Tinkercad :
https://www.tinkercad.com/things/7QK3tCraoTB